# Improved performance of function lp